+++ src/App.jsx
@@ -1,58 +1,38 @@
 import * as React from 'react';
+import TodoList from './TodoList.jsx';
+import AddTodoForm from './AddTodoForm.jsx'; 
 
-/*
-const welcome = {
-  greeting: 'Hey',
-  title: 'React',
-}
-*/
 
-/*function getTitle(title) {
-  return title;
-}*/
+// const title = 'React';
 
-const title = 'React';
 
-const todoList = [
-  {
-    id: 1,
-    title: 'Complete assignment',
-    date: '2020-02-15',
-  },
-  {
-    id: 2,
-    title: 'Submit assignment',
-    date: '2022-10-30',
-  },
-  {
-    id: 3,
-    title: 'Check assignment',
-    date: '2024-08-01',
-  },
-];
 
 
 function App() {
   return (
     <div>
-      {/*<h1>
-        {welcome.greeting} {welcome.title}
-      </h1>*/}
       <h1>Todo List</h1>
-
-      <label htmlFor="search">Search: </label>
-      <input id="search" type="text" />
+      <AddTodoForm />
+      <TodoList />
+      <Search />
+  
 
       <hr />
-      <ul>
-        {todoList.map(function (item) {
-          return (<> 
-        <li key={item.id}>{item.title}</li>
-          </>)
-        })}
-      </ul>
+        
+
     </div>
   );
+}
+
+
+
+function Search() {
+  return(
+    <div>
+      <label htmlFor="search">Search: </label>
+      <input id="search" type="text" />
+    </div>
+  )
 }
 
 export default App;
